Feat: update user pages & component classification
