+++ src/common/Popular.jsx
@@ -1,5 +1,5 @@
-import styles from "../../pages/Home/HomeStyle.module.scss";
-import Follower from "./Follower";
+import styles from "../pages/HomeStyle.module.scss";
+import Follower from "../components/Home/Follower";
 
 const Popular = () => {
   return (
