+++ src/components/Home/Follower.jsx
@@ -1,7 +1,7 @@
 import { ReactComponent as FollowButton} from "../../assets/image/FollowButton.svg";
 // import { ReactComponent as StopFollow } from "../../assets/image/StopFollow.svg";
 
-import styles from "../../pages/Home/HomeStyle.module.scss";
+import styles from "../../pages/HomeStyle.module.scss";
 const Follower = () => {
   return (
     <div className={styles.followerContainer}>
