+++ src/pages/UserProfilePage.jsx
@@ -1,14 +1,13 @@
-import styles from "../pages/Home/HomeStyle.module.scss";
-import NavBar from "../components/Navbar/HomeNav/NavBar";
-import Popular from "../components/Home/Popular";
+import styles from "../pages/HomeStyle.module.scss";
+import UserNav from '../components/User/UserNav'
+import Popular from "../common/Popular";
 import UserMain from "../components/User/UserMain";
-// import TweetModal from "./TweetModal";
 
 const UserProfilePage = () => {
   return (
     <div className={styles.homeContainer}>
       <div className={styles.mainContainer}>
-        <NavBar />
+        <UserNav />
         <UserMain />
         <Popular />
       </div>
